swap out index files
