+++ checkapi/index.js
@@ -1,55 +1,61 @@
 'use strict'
 
 require('dotenv').config()
-const appInsights = require('applicationinsights')
-const express = require('express')
-// const compression = require('compression')
-const morgan = require('morgan')
-const bodyParser = require('body-parser')
-const helmet = require('helmet')
-const errorHandler = require('errorhandler')
-// controllers
-const ping = require('./controllers/ping')
-const { completeCheck } = require('./controllers/complete-check')
-const { auth } = require('./controllers/auth')
-const storageService = require('./services/data-access/storage.service')
-const config = require('./config')
+const http = require('http')
+// const timer = require('simple-timer')
+const fs = require('fs')
+const path = require('path')
+const azure = require('azure-storage')
+const queueService = azure.createQueueService()
+const queueName = 'completedchecks'
 
-// initialise monitoring
-if (config.APPINSIGHTS_INSTRUMENTATIONKEY) {
-  appInsights.setup()
-  appInsights.start()
-}
-
-storageService.init()
-
-const app = express()
-
-// configure express defaults
-app.set('port', process.env.PORT || 3003)
-// app.use(compression())
-if (process.env.NODE_ENV !== 'production') {
-  app.use(morgan('dev'))
-  app.use(errorHandler())
-}
-app.use(bodyParser.json())
-app.disable('x-powered-by')
-app.use(helmet())
-app.use(helmet.contentSecurityPolicy({
-  directives: {
-    defaultSrc: ["'none'"]
+queueService.createQueueIfNotExists(queueName, function (error, result, response) {
+  if (error) {
+    throw new Error('unable to connect to azure storage:' + error.message)
   }
-}))
-
-app.use(express.static('public'))
-// routes
-app.post('/complete-check', completeCheck)
-app.post('/auth', auth)
-app.get('/ping', ping.get)
-
-app.listen(app.get('port'), () => {
-  console.log(('  App is running on port %d in %s mode'), app.get('port'), app.get('env'))
-  console.log('  Press CTRL-C to stop\n')
+  console.log('connected to azure queue service')
 })
 
-module.exports = app
+http.createServer((request, response) => {
+  // const { headers, method, url } = request
+  if (request.method === 'GET' && request.url === '/') {
+    response.write('Raw')
+    response.statusCode = 200
+    response.end()
+  }
+
+  if (request.method === 'GET' && request.url === '/loaderio-99a9565f58325293e217a36aa3ae695d.txt') {
+    response.writeHead(200)
+    response.end(fs.readFileSync(path.join(__dirname, 'public', 'loaderio-99a9565f58325293e217a36aa3ae695d.txt')))
+  }
+
+  if (request.method !== 'POST' || request.url !== '/complete-check') {
+    response.statusCode = 404
+    response.end()
+  }
+  let body = []
+  request.on('error', (err) => {
+    console.error(err)
+  }).on('data', (chunk) => {
+    body.push(chunk)
+  }).on('end', () => {
+    body = Buffer.concat(body).toString()
+    response.on('error', (err) => {
+      console.error(err)
+    })
+
+    // TODO validation of request, authentication
+    // timer.start('request')
+    queueService.createMessage(queueName, body, function (error, result, queueResponse) {
+      // timer.stop('request')
+      // console.log(`queue call took ${timer.get('request').delta}ms`)
+      if (error) {
+        console.error(error)
+        response.writeHead(500)
+        response.end()
+      }
+      response.writeHead(200, { 'Content-Type': 'application/json' })
+      response.end()
+    })
+  })
+}).listen(process.env.PORT || 3003)
